# packages/sdk/openapi/agentflow.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentFlow API",
    "version": "0.1.0",
    "description": "REST API for AgentFlow — an open-source workflow automation platform.\n\n**Authentication** — send `Authorization: Bearer <token>` with a JWT access token (POST /api/auth/login) or a personal API key (`af_…`, POST /api/api-keys).\n\n**Rate limiting** — 200 requests/minute per user or IP by default; stricter limits apply to auth, AI and checkout endpoints.\n\n**Pagination** — list endpoints accept `page` and `limit` (max 100) and return the applied values in the `X-Page` and `X-Limit` response headers.",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "List workflows of the active organization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            },
            "required": true,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "example": 25
            },
            "required": true,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflows, most recently updated first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workflow"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Create a workflow",
        "description": "Enforces the workflow limit of the organization's plan (403 WORKFLOW_LIMIT_REACHED).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Workflow metadata",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workflow"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or user has no organization (NO_ORG)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Plan workflow limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflows/{id}/run": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Run a workflow manually",
        "description": "Enqueues an execution (BullMQ when Redis is available, inline otherwise) and records usage against the organization's monthly quota.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "202": {
            "description": "Execution created and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Monthly execution quota exceeded (QUOTA_EXCEEDED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "NOT_FOUND"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Field-level validation errors (VALIDATION_ERROR only)"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "description": "Error envelope used by every non-2xx response"
      },
...
      "Workflow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx4wf777"
          },
          "name": {
            "type": "string",
            "example": "Sync orders to Sheets"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "ARCHIVED"
            ],
            "example": "ACTIVE"
          },
          "orgId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "nodes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "node-1"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "trigger",
                    "action",
                    "logic",
                    "advanced"
                  ],
                  "description": "Canvas category"
                },
                "position": {
                  "type": "object",
                  "properties": {
                    "x": {
                      "type": "number"
                    },
                    "y": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "x",
                    "y"
                  ]
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Executable node type (webhook, http, condition, …)",
                      "example": "http"
                    },
                    "label": {
                      "type": "string",
                      "example": "HTTP Request"
                    },
                    "description": {
                      "type": "string"
                    },
                    "config": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "type",
                    "label",
                    "description",
                    "config"
                  ]
                }
              },
              "required": [
                "id",
                "type",
                "position",
                "data"
              ]
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "edge-1"
                },
                "source": {
                  "type": "string",
                  "description": "Source node id"
                },
                "target": {
                  "type": "string",
                  "description": "Target node id"
                },
                "sourceHandle": {
                  "type": "string"
                },
                "targetHandle": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "condition": {},
                "animated": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "source",
                "target",
                "animated"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "orgId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AgentFlow access token (from POST /api/auth/login) or a personal API key (`af_…`, from POST /api/api-keys). Send as `Authorization: Bearer <token>`."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Signature",
        "description": "HMAC-SHA256 hex digest of the raw request body, computed with the webhook secret (optional `sha256=` prefix). Required on every /api/webhooks/trigger/{path} call."
      },
      "stripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature header (t=…,v1=…)."
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    },
    {
      "url": "https://api.agentflow.dev",
      "description": "Production (placeholder)"
    }
  ],
...
}